+++ apps/base-docs/docs/tools/oracles.md
@@ -72,7 +72,7 @@ To use Chainlink datafeeds, you may need [LINK](https://docs.chain.link/resource
 
 [Chronicle](https://chroniclelabs.org/) provides a number of [Oracles](https://chroniclelabs.org/dashboard) for Base.
 
-See [this guide](https://docs.chroniclelabs.org/Developers/tutorials/Remix) to learn how to use the Chronicle Oracles.
+See [this guide](https://docs.chroniclelabs.org/Builders/tutorials/Remix) to learn how to use the Chronicle Oracles.
 
 #### Supported Networks
