+++ apps/web/src/components/Builders/MiniKit/bg1.png
