+++ apps/web/src/components/Builders/MiniKit/bg2.png
