+++ apps/web/src/components/Builders/Shared/AnimatedBeam.tsx
